refactor & rename to svg-spinner
