+++ apps/builder/app/dashboard/loading-state/svg-spinner.ts
